Add hosts not showing up troubleshooting modal for infra env table
